implement actions for head of department

+++ DigitNow.Microservice.DocumentManagement/DigitNow.Domain.DocumentManagement/Public/IncomingDocuments/Models/CreateWorkflowDecisionRequest.cs
@@ -9,7 +9,7 @@ namespace DigitNow.Domain.DocumentManagement.Public.IncomingDocuments.Models
         public int ActionType { get; set; }
         public string? Remarks { get; set; }
         public string? DeclineReason { get; set; }
-        public int? RecipientId { get; set; }
+        public long? RecipientId { get; set; }
         public int? Resolution { get; set; }
         public int? Decision { get; set; }
         public DateTime? OpionionRequestedUntil { get; set; }
